+++ CHANGELOG.md
@@ -19,6 +19,8 @@
 
 * [HexaPDF::Document::Destinations::Destination::new] to also accept a hash
 * [HexaPDF::Type::Catalog] auto-conversion of /Outlines to correct class
+* [HexaPDF::Type::AcroForm::Form#flatten] to return the unflattened form fields
+  instead of the widgets
 
 
 ## 0.26.2 - 2022-10-22